+++ layer2/rendering/__init__.py
@@ -1,6 +1,3 @@
 # flake8: noqa
-from .asset_container_card import CARD_ASSET_REF
-from .asset_container_iso import ISO_ASSET_REF
-from .asset_container_ui import UI_ASSET_REF
 from .rendering_processor import RENDER_PROC_REF
 from .utils import UIElemSprite, UIElemType
